Add hf model custom components support
